Changes to unit tests to work on python 3.12
