+++ pyproject.toml
@@ -148,6 +148,10 @@ filterwarnings = [
     # Our dependency is via opentelemetry-exporter-otlp==1.27 which depends opentelemetry-proto==1.27
     # which depends protobuf>=3.19, <5.0
     # https://github.com/open-telemetry/opentelemetry-python/pull/4206/commits addresses this but is not yet in a release
+    "ignore:.*custom tp_new.*:DeprecationWarning",
+    # Ignore warning about deprecated throw() call https://github.com/bluesky/bluesky/issues/1817
+    "ignore:.*signature of throw\\(\\) is deprecated.*:DeprecationWarning",
+
 ]
 # Doctest python code in docs, python code in src docstrings, test functions in tests
 testpaths = "docs src tests/unit_tests"
