+++ tests/unit_tests/hyperion/device_setup_plans/test_utils.py
@@ -4,6 +4,7 @@ import pytest
 from bluesky import plan_stubs as bps
 from bluesky.utils import FailedStatus
 from dodal.beamlines import i03
+from ophyd_async.core import get_mock_put
 from ophyd.status import Status
 
 from mx_bluesky.hyperion.device_setup_plans.utils import (
@@ -25,14 +26,13 @@ class MyTestException(Exception):
 
 
 def test_given_plan_raises_when_exception_raised_then_eiger_disarmed_and_correct_exception_returned(
-    mock_eiger, RE
+    mock_eiger, detector_motion, RE
 ):
     def my_plan():
         yield from bps.null()
         raise MyTestException()
 
     eiger = mock_eiger
-    detector_motion = MagicMock()
 
     with pytest.raises(MyTestException):
         RE(
@@ -74,9 +74,8 @@ def test_given_shutter_open_fails_then_eiger_disarmed_and_correct_exception_retu
 
 
 def test_given_detector_move_fails_then_eiger_disarmed_and_correct_exception_returned(
-    mock_eiger, null_plan, RE
+    mock_eiger, detector_motion, null_plan, RE
 ):
-    detector_motion = MagicMock()
     status = Status()
     status.set_exception(MyTestException())
     detector_motion.shutter.set = MagicMock(return_value=status)
@@ -90,5 +89,5 @@ def test_given_detector_move_fails_then_eiger_disarmed_and_correct_exception_ret
     assert e.value.args[0] is status
 
     mock_eiger.async_stage.assert_called_once()
-    detector_motion.z.set.assert_called_once()
+    get_mock_put(detector_motion.z.user_setpoint).assert_called_once()
     mock_eiger.disarm_detector.assert_called_once()
