+++ tests/unit_tests/hyperion/experiment_plans/test_pin_centre_then_xray_centre_plan.py
@@ -9,6 +9,7 @@ from dodal.devices.backlight import Backlight, BacklightPosition
 from dodal.devices.detector.detector_motion import ShutterState
 from dodal.devices.smargon import Smargon
 from dodal.devices.synchrotron import SynchrotronMode
+from mx_bluesky.hyperion.experiment_plans.grid_detect_then_xray_centre_plan import GridDetectThenXRayCentreComposite
 
 from mx_bluesky.hyperion.experiment_plans.pin_centre_then_xray_centre_plan import (
     create_parameters_for_grid_detection,
@@ -51,12 +52,13 @@ def test_when_pin_centre_xray_centre_called_then_plan_runs_correctly(
     mock_detect_and_do_gridscan: MagicMock,
     mock_pin_tip_centre: MagicMock,
     test_pin_centre_then_xray_centre_params: PinTipCentreThenXrayCentre,
+    grid_detect_devices: GridDetectThenXRayCentreComposite,
     test_config_files,
 ):
     RE = RunEngine()
     RE(
         pin_centre_then_xray_centre_plan(
-            MagicMock(), test_pin_centre_then_xray_centre_params, test_config_files
+            grid_detect_devices, test_pin_centre_then_xray_centre_params, test_config_files
         )
     )
 
@@ -169,6 +171,7 @@ def test_pin_centre_then_xray_centre_plan_activates_ispyb_callback_before_pin_ti
     mock_pin_tip_centre_plan,
     sim_run_engine: RunEngineSimulator,
     test_pin_centre_then_xray_centre_params: PinTipCentreThenXrayCentre,
+    grid_detect_devices: GridDetectThenXRayCentreComposite,
     test_config_files,
 ):
     mock_detect_grid_and_do_gridscan.return_value = iter(
@@ -178,7 +181,7 @@ def test_pin_centre_then_xray_centre_plan_activates_ispyb_callback_before_pin_ti
 
     msgs = sim_run_engine.simulate_plan(
         pin_centre_then_xray_centre_plan(
-            MagicMock(),
+            grid_detect_devices,
             test_pin_centre_then_xray_centre_params,
             test_config_files["oav_config_json"],
         )
@@ -209,8 +212,7 @@ def test_pin_centre_then_xray_centre_plan_activates_ispyb_callback_before_pin_ti
 def test_pin_centre_then_xray_centre_plan_sets_up_backlight_and_aperture(
     mock_detect_grid_and_do_gridscan,
     mock_pin_tip_centre_plan,
-    backlight: Backlight,
-    aperture_scatterguard: ApertureScatterguard,
+    grid_detect_devices: GridDetectThenXRayCentreComposite,
     sim_run_engine: RunEngineSimulator,
     test_pin_centre_then_xray_centre_params: PinTipCentreThenXrayCentre,
     test_config_files,
@@ -220,13 +222,9 @@ def test_pin_centre_then_xray_centre_plan_sets_up_backlight_and_aperture(
     )
     mock_pin_tip_centre_plan.return_value = iter([Msg("pin_tip_centre_plan")])
 
-    mock_composite = MagicMock()
-    mock_composite.aperture_scatterguard = aperture_scatterguard
-    mock_composite.backlight = backlight
-
     msgs = sim_run_engine.simulate_plan(
         pin_centre_then_xray_centre_plan(
-            mock_composite,
+            grid_detect_devices,
             test_pin_centre_then_xray_centre_params,
             test_config_files["oav_config_json"],
         )
